fix(exa): provision usage partitions for tests
