+++ apps/web/src/tests/setup/workerSetup.ts
@@ -6,6 +6,7 @@ import { LEGACY_KILOCLAW_PRICE_VERSION } from '@kilocode/db';
 import { kiloclaw_subscriptions } from '@kilocode/db/schema';
 import { drizzle } from 'drizzle-orm/node-postgres';
 import { migrate } from 'drizzle-orm/node-postgres/migrator';
+import { provisionExaUsageLogPartitions } from '@/lib/exa-usage-partitions';
 import { existsSync, writeFileSync, mkdirSync } from 'fs';
 import { join } from 'path';
 import { shutdownPosthog } from '@/lib/posthog';
@@ -51,6 +52,17 @@ beforeAll(async () => {
   try {
     const testDb = drizzle(testPool);
     await migrate(testDb, { migrationsFolder: '../../packages/db/src/migrations' });
+
+    // Production keeps this rolling window current via cron. Each Jest worker
+    // starts from the static migration snapshot, so it needs the same window
+    // before tests insert rows whose created_at defaults to now().
+    const { errors: partitionErrors } = await provisionExaUsageLogPartitions(testDb);
+    if (partitionErrors.length > 0) {
+      const [{ name, error }] = partitionErrors;
+      throw new Error(
+        `Failed to create Exa usage log partition ${name}: ${error instanceof Error ? error.message : String(error)}`
+      );
+    }
   } finally {
     await testPool.end();
   }
